requirements.txt
google-genai
requests>=2.32.3
html2text>=2024.2.26
rich>=13.9.4
ollama>=0.4.7
chromadb>=0.6.3
geopy>=2.4.1
mcp; python_version >= "3.10"
Pillow>=10.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
websockets>=14.1